Telegram notifications (#54)
* added orders anomalies summary worker

* linted

* linted

* fixed with mypy

* added tests

* updated .development.env

* Telegram client

* fix

* fix

* fix

* fixex mypy

* fix mypy

* fixed mypy

* fixed mypy

* fix

* fix mypy

* fix

* code refactor

* fixed bugs

* Added messenger for summary worker, mirror fixes

* Handling for 0 values

* .

* lint

* Middle commit

* Ref

* Implemented telegram client in workers, fixed tests, refactor due to mypy, added migration

* Returned python version

* update lock file

* Removed unused imports

* Removed tg secret, some mirror fixes, renamed to stubs. Didn't found special stub dependency for discord-webhook

* Env details for PR

* add

* changed replicas

* .

---------

+++ app/db/models/volume.py
@@ -1,6 +1,7 @@
 from uuid import UUID
 
-from sqlalchemy import Float, ForeignKey, Integer
+from _decimal import Decimal
+from sqlalchemy import DECIMAL, ForeignKey, Integer
 from sqlalchemy.dialects.postgresql import UUID as pg_UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -10,7 +11,7 @@ from app.common.database import BaseModel
 class Volume(BaseModel):
     __tablename__ = "volumes"
 
-    bid_ask_ratio: Mapped[float] = mapped_column(Float, nullable=False)
+    bid_ask_ratio: Mapped[Decimal] = mapped_column(DECIMAL, nullable=False)
     average_volume: Mapped[int] = mapped_column(Integer, nullable=False)
     launch_id: Mapped[UUID] = mapped_column(
         pg_UUID, nullable=False, index=True
